refactor(#30) : 도메인 재분류
- category, topic, hashtag 관련 aggregate 를 search domain 으로 분류
- servlet 을 Controller 패키지로 이동
