--- src/main/java/com/blog/tech/domain/hashtag/controller/HashtagController.java
@@ -1,4 +1,0 @@
-package com.blog.tech.domain.hashtag.controller;
-
-public class HashtagController {
-}
