--- src/main/java/com/blog/tech/domain/hashtag/entity/ConnectHashtag.java
@@ -1,16 +1,0 @@
-package com.blog.tech.domain.hashtag.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class ConnectHashtag {
-
-	private Long id;
-	private Long hashtagId;
-	private Long postId;
-
-}
