--- src/main/java/com/blog/tech/domain/hashtag/entity/Hashtag.java
@@ -1,15 +1,0 @@
-package com.blog.tech.domain.hashtag.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class Hashtag {
-
-	private Long id;
-	private String name;
-
-}
