--- src/main/java/com/blog/tech/domain/hashtag/repository/ifs/ConnectHashtagRepository.java
@@ -1,7 +1,0 @@
-package com.blog.tech.domain.hashtag.repository.ifs;
-
-import com.blog.tech.domain.common.CRUDIfs;
-import com.blog.tech.domain.hashtag.entity.ConnectHashtag;
-
-public interface ConnectHashtagRepository extends CRUDIfs<ConnectHashtag> {
-}
