--- src/main/java/com/blog/tech/domain/hashtag/repository/ifs/HashtagRepository.java
@@ -1,7 +1,0 @@
-package com.blog.tech.domain.hashtag.repository.ifs;
-
-import com.blog.tech.domain.common.CRUDIfs;
-import com.blog.tech.domain.hashtag.entity.Hashtag;
-
-public interface HashtagRepository extends CRUDIfs<Hashtag> {
-}
