+++ src/main/java/com/blog/tech/domain/post/controller/PostController.java
@@ -1,10 +1,8 @@
 package com.blog.tech.domain.post.controller;
 
 import java.sql.SQLException;
-import java.util.List;
 
 import com.blog.tech.domain.post.dto.request.PostRequestBean;
-import com.blog.tech.domain.post.dto.response.CategoryResponseBean;
 import com.blog.tech.domain.post.service.PostService;
 
 public class PostController {
@@ -15,11 +13,8 @@ public class PostController {
 		this.postService = postService;
 	}
 
-	public void writeOnPost(final PostRequestBean request) throws SQLException {
+	public void writeOnPost(final PostRequestBean request) throws SQLException, SQLException {
 		postService.writeOnPost(request);
 	}
 
-	public List<CategoryResponseBean> getAllCategories() throws SQLException {
-		return postService.getAllCategories();
-	}
 }
