--- src/main/java/com/blog/tech/domain/post/entity/Category.java
@@ -1,17 +1,0 @@
-package com.blog.tech.domain.post.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@Builder
-@AllArgsConstructor
-public class Category {
-
-	private Long id;
-	private String name;
-
-}
