--- src/main/java/com/blog/tech/domain/post/entity/Topic.java
@@ -1,18 +1,0 @@
-package com.blog.tech.domain.post.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@Builder
-@AllArgsConstructor
-public class Topic {
-
-	private Long id;
-	private Long categoryId;
-	private String name;
-
-}
