--- src/main/java/com/blog/tech/domain/post/repository/ifs/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.blog.tech.domain.post.repository.ifs;
-
-import com.blog.tech.domain.common.CRUDIfs;
-import com.blog.tech.domain.post.entity.Category;
-
-public interface CategoryRepository extends CRUDIfs<Category> {
-}
