+++ src/main/java/com/blog/tech/domain/search/entity/Category.java
@@ -1,0 +1,17 @@
+package com.blog.tech.domain.search.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+public class Category {
+
+	private Long id;
+	private String name;
+
+}
