+++ src/main/java/com/blog/tech/domain/search/entity/ConnectHashtag.java
@@ -1,0 +1,16 @@
+package com.blog.tech.domain.search.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class ConnectHashtag {
+
+	private Long id;
+	private Long hashtagId;
+	private Long postId;
+
+}
