+++ src/main/java/com/blog/tech/domain/search/entity/Hashtag.java
@@ -1,0 +1,15 @@
+package com.blog.tech.domain.search.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class Hashtag {
+
+	private Long id;
+	private String name;
+
+}
