+++ src/main/java/com/blog/tech/domain/search/entity/Topic.java
@@ -1,0 +1,18 @@
+package com.blog.tech.domain.search.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+public class Topic {
+
+	private Long id;
+	private Long categoryId;
+	private String name;
+
+}
