+++ src/main/java/com/blog/tech/domain/search/repository/ifs/CategoryRepository.java
@@ -1,0 +1,7 @@
+package com.blog.tech.domain.search.repository.ifs;
+
+import com.blog.tech.domain.common.CRUDIfs;
+import com.blog.tech.domain.search.entity.Category;
+
+public interface CategoryRepository extends CRUDIfs<Category> {
+}
