+++ src/main/java/com/blog/tech/global/config/ControllerLoaderListener.java
@@ -7,13 +7,16 @@ import com.blog.tech.domain.member.repository.dao.MemberDao;
 import com.blog.tech.domain.member.repository.dao.MemberInfoDao;
 import com.blog.tech.domain.member.service.MemberService;
 import com.blog.tech.domain.post.controller.PostController;
-import com.blog.tech.domain.post.repository.dao.CategoryDao;
+import com.blog.tech.domain.search.controller.SearchController;
+import com.blog.tech.domain.search.repository.dao.CategoryDao;
 import com.blog.tech.domain.post.repository.dao.CommentDao;
-import com.blog.tech.domain.post.repository.dao.PostAttachDao;
 import com.blog.tech.domain.post.repository.dao.PostDao;
 import com.blog.tech.domain.post.repository.dao.ReplyDao;
-import com.blog.tech.domain.post.repository.dao.TopicDao;
+import com.blog.tech.domain.search.repository.dao.ConnectHashtagDao;
+import com.blog.tech.domain.search.repository.dao.HashtagDao;
+import com.blog.tech.domain.search.repository.dao.TopicDao;
 import com.blog.tech.domain.post.service.PostService;
+import com.blog.tech.domain.search.service.SearchService;
 
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
@@ -25,18 +28,26 @@ public class ControllerLoaderListener implements ServletContextListener {
 		final Connection conn = (Connection)sce.getServletContext().getAttribute("conn");
 		setMemberController(conn, sce);
 		setPostController(conn, sce);
+		setSearchController(conn, sce);
 	}
 
 	private void setPostController(final Connection conn, final ServletContextEvent sce) {
-		final CategoryDao category = new CategoryDao(conn);
-		final CommentDao comment = new CommentDao(conn);
-		final PostDao post = new PostDao(conn);
-		final ReplyDao reply = new ReplyDao(conn);
-		final TopicDao topic = new TopicDao(conn);
-		final PostAttachDao postAttach = new PostAttachDao(conn);
-		final PostService postService = new PostService(category, post, topic, reply, comment, postAttach);
+		final CommentDao commentDao = new CommentDao(conn);
+		final PostDao postDao = new PostDao(conn);
+		final ReplyDao replyDao = new ReplyDao(conn);
+		final PostService postService = new PostService(postDao, replyDao, commentDao);
 		final PostController postController = new PostController(postService);
 		sce.getServletContext().setAttribute("postController", postController);
+	}
+
+	private void setSearchController(final Connection conn, final ServletContextEvent sce) {
+		final CategoryDao category = new CategoryDao(conn);
+		final TopicDao topic = new TopicDao(conn);
+		final HashtagDao hashtag = new HashtagDao(conn);
+		final ConnectHashtagDao connect = new ConnectHashtagDao(conn);
+		final SearchService searchService = new SearchService(hashtag, connect, category, topic);
+		final SearchController searchController = new SearchController(searchService);
+		sce.getServletContext().setAttribute("searchController", searchController);
 	}
 
 	private void setMemberController(final Connection conn, final ServletContextEvent sce) {
